Refactor code using solid principles
