+++ introduction.rb
@@ -1,0 +1,10 @@
+module Introductions
+    def welcome_message
+    puts 'Welcome to School Library App!'
+    end
+
+    def farewell_message
+        exit_message
+        exit
+    end
+end
