+++ options.rb
@@ -1,0 +1,20 @@
+module Options   
+def select_option(selected, app)
+    case selected
+    when 1
+      app.list_all_books
+    when 2
+      app.list_all_people
+    when 3
+      app.create_person
+    when 4
+      app.create_book
+    when 5
+      app.create_rental
+    when 6
+      app.list_rentals_for_person
+    else
+        show_invalid_input
+    end
+  end
+end
